+++ tests/unit/Requests/FiltersTest.php
@@ -43,4 +43,32 @@ class FiltersTest extends TestCase
         $this->filters->add('gender', 'Attack Helicopter');
         $this->assertEquals($expected, $this->filters->toArray());
     }
+
+    public function test_it_handles_empty_strings_correctly()
+    {
+        $expected = [
+            'name' => ['Example name'],
+            'city' => ['Indianapolis', 'Des Moines'],
+            'gender' => [],
+            'age' => [],
+        ];
+
+        $this->filters->add('gender', '');
+        $this->filters->add('age', '  ');
+        $this->assertEquals($expected, $this->filters->toArray());
+    }
+
+    public function test_it_filters_out_nested_empty_strings()
+    {
+        $expected = [
+            'name' => ['Example name'],
+            'city' => ['Indianapolis', 'Des Moines'],
+            'gender' => [],
+            'age' => [],
+        ];
+
+        $this->filters->add('gender', ['']);
+        $this->filters->add('age', [' ', '     ']);
+        $this->assertEquals($expected, $this->filters->toArray());
+    }
 }